Movement and player added
I created a player movement script and created a temp player as well as a platform for him to move on. Player movement is basic but jumping requires a ground check, dont move or change the ground check or it would mess up the movement. Feel free to look at the code